worker/requirements.txt
celery
redis
psycopg2-binary
minio
google-genai
pydantic
python-dotenv